test: fix flaky vinyl/gh-3395-read-prepared-uncommitted.test.lua
The test fails on aarch64 CI runners with errors like this:

```
[113] vinyl/gh-3395-read-prepared-uncommitted.test.l>               [ fail ]
[113]
[113] Test failed! Result content mismatch:
[113] --- vinyl/gh-3395-read-prepared-uncommitted.result	Wed Feb 12 10:13:31 2025
[113] +++ /tmp/t/rejects/vinyl/gh-3395-read-prepared-uncommitted.reject	Wed Feb 12 10:24:09 2025
[113] @@ -119,17 +119,18 @@
[113]  --
[113]  read_prepared_with_delay(false)
[113]   | ---
[113] - | - - [2, 2]
[113] - |   - [3, 2]
[113] - | ...
[113] --- 2. Tuple is not rolled back so it is visible to all transactions.
[113] ---
[113] -read_prepared_with_delay(true)
[113] - | ---
[113]   | - - [1, 2]
[113]   |   - [2, 2]
[113]   |   - [3, 2]
[113]   | ...
[113] +-- 2. Tuple is not rolled back so it is visible to all transactions.
[113] +--
[113] +read_prepared_with_delay(true)
[113] + | ---
[113] + | - - [1, 2]
[113] + |   - [2, 2]
[113] + |   - [3, 2]
[113] + | ...
[113]
[113]  -- Give WAL thread time to catch up.
[113]  --
[113] @@ -139,7 +140,8 @@
[113]
[113]  sk:select{2}
[113]   | ---
[113] - | - - [2, 2]
[113] + | - - [1, 2]
[113] + |   - [2, 2]
[113]   |   - [3, 2]
[113]   | ...
```

Here's why it happens. Note the sleep between clearing ERRINJ_WAL_DELAY
and ERRINJ_VY_READ_PAGE_DELAY injections in read_prepared_with_delay.
Usually, it should be enough for the WAL rollback to complete but it
looks like on aarch64 it isn't. As a result, the reader returns
a prepared tuple before it's rolled back. Instead of sleeping, let's
join the writer fiber to make sure the rollback is complete before
the reader resumes.

The patch isn't as simple as that because of the way the flag
is_tx_faster_than_wal is handled in read_prepared_with_delay.
It emulates the situation when the reader actually manages to read
a prepared tuple because of a slow WAL with another error injection
ERRINJ_RELAY_FASTER_THAN_TX, which prevents the writer fiber from
returning to the TX thread after receiving a WAL error, so we can't
just join it. Actually, using this error injection seems pointless.
Instead we could clear ERRINJ_WAL_DELAY after finishing the read,
and this is what this patch does.

A couple more notes:
 - Drop the channels used for starting a read fiber and getting the read
   result. We can get the result with fiber.join() while stalling the
   fiber at startup seems pointless - a yield should be enough.
 - Set the defer_deletes flag on the test space to make sure the writer
   doesn't yield on disk read, otherwise the test may hang. It didn't
   hang already, because the disk read was luckily reflected by the
   bloom filter. BTW deferred DELETEs were enabled for all spaces when
   the test was introduced.

Closes tarantool/tarantool-qa#202

NO_DOC=testing
NO_CHANGELOG=testing

(cherry picked from commit b8018147e73e213a1cfc60cdda98145dbf9e892c)