test: fix flaky vinyl/errinj.test.lua
There are two issues with this test.

The first one is an attempt to make a dead fiber joinable:

```
[039] vinyl/errinj.test.lua                                         [ fail ]
[039]
[039] Test failed! Result content mismatch:
[039] --- vinyl/errinj.result	Thu Jan 30 15:14:04 2025
[039] +++ /tmp/t/rejects/vinyl/errinj.reject	Thu Jan 30 15:20:26 2025
[039] @@ -712,6 +712,7 @@
[039]  ...
[039]  f:set_joinable(true);
[039]  ---
[039] +- error: the fiber is dead
[039]  ...
[039]  ok, err = nil;
[039]  ---
[039] @@ -722,6 +723,8 @@
[039]      ok, err = f:join(0.01)
[039]  until ok;
[039]  ---
[039] +- error: '[string "repeat     box.snapshot()     errinj.set("ERR..."]:1: the fiber
[039] +    is dead'
[039]  ...
[039]  test_run:cmd("setopt delimiter ''");
[039]  ---
[039]
[039] [test-run server "vinyl"] Last 15 lines of the log file /tmp/t/039_vinyl/vinyl.log:
[039]  to --- {1: 970}
[039] ...
[039]
[039] 2025-01-30 15:20:26.164 [58596] main/131/console/unix/:/test_run I> wait vclock: --- {1: 969}
[039] ...
[039]  to --- {1: 970}
[039] ...
[039]
[039] 2025-01-30 15:20:26.173 [58596] main/131/console/unix/:/test_run I> wait vclock: --- {1: 969}
[039] ...
[039]  to --- {1: 970}
[039] ...
```

Apparently, this happens because there's a yield between creating
a fiber and making it joinable, which opens a time window for the
fiber to complete. Fix the issue by eliminating the yield.

The second issue is an attempt to cancel a dead fiber:

```
[012] Test failed! Result content mismatch:
[012] --- vinyl/errinj.result	Mon Feb 17 14:42:17 2025
[012] +++ /tmp/t/rejects/vinyl/errinj.reject	Mon Feb 17 14:46:05 2025
[012] @@ -220,6 +220,7 @@
[012]  ...
[012]  fiber.cancel(f1)
[012]  ---
[012] +- error: the fiber is dead
[012]  ...
[012]  fiber.sleep(0.1)
[012]  ---
```

This issue is caused by the read page timeout being too small -
the fiber manages to complete before it's cancelled. Let's use
the ERRINJ_VY_READ_PAGE_DELAY instead of ERRINJ_VY_READ_PAGE_TIMEOUT
to make sure it doesn't happen. Also, let's join the cancelled fibers
to check their exit statuses.

NO_DOC=testing
NO_CHANGELOG=testing