memtx: fix a crash caused by mhash misusage in MVCC
We use mhash in memtx MVCC to store trackers of reads that have read
nothing, we call them point holes. When handling a write to such hole,
the tracker should be deleted because we can use the new tuple to store
reads instead. Deletion flow in mhash is: we find bucket of the element
with `find`, then we free the bucket with `del`. It can seem that the
element is not needed on `del` because bucket id is used. However, it
can be used on incremental resize of mhash. And, since we delete the
point holes before releasing the bucket in mhash, in the rare case of
incremental resize Tarantool will be aborted by mhash consistency check.
The commit fixes the problem - simply release the bucket before deleting
the object.

Along the way, the commit fixes another misusage that doesn't actually
break something. Method `put_slot` is internal so shouldn't be used
manually - let's use more convenient `get` instead.

Closes #11022

NO_DOC=bugfix