memtx: use functional key for functional indexes in mvcc
Currently, there are several places where we compare tuples with other
tuples, keys, or even calculate their hashes. The problem is if the
index is functional we should use the functional key instead of original
tuple for such operations. Moreover, an attempt to call `tuple_hash`
with functional index definition will fail with an assertion.

Firstly, in order to be able to call Lua function inside read methods of
a functional index, the commit replaces them with LuaC implementations.

There is no mapping from tuple to its functional key for a particular
index, so the commit adds one to MVCC - it stores functional keys of
dirty tuples (note that amount of dirty tuples is far less than amount
of all tuples). However, we can't just obtain all functional keys for a
tuple when it becomes dirty - it can happen during a read request to
another index, so we can be inside of an FFI call, so we are not allowed
to call another Lua function to obtain the functional key. So let's just
fill the mapping lazily. Moreover, in this case performance of requests to
non-functional indexes won't be affected at all.

Note that we could simply call the function of the index to obtain key
of a tuple every time we need it, without any mapping. But such approach
could dramatically slow things down. For example, if we used `count`
often, we would need to call the function for every story on every count!

Closes #11100

NO_CHANGELOG=later
NO_DOC=later