index: move iterator type check to common code
Currently, compatibility of the iterator type with the target index is
performed by the index implementation. This results in code duplication,
especially, in the scope of PP/NP iterator support in vinyl. Let's move
this check to the common code. We already have the key_validate() helper
function, which is called before opening an iterator and passed all
iteration parameters. Let's perform the iterator type check there as
well and rename it to iterator_type().

Needed for #10647

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring