index: move PP/NP iterator preparation to common code
This patch moves the prepare_start_prefix_iterator() function from
memtx_tree.cc to index.h/cc so that it can be used in other engines.
The only change done to the function along the way is replacing the
memtx_tree_key_data argument with a raw key so that now the key hint
is supposed to be computed by the caller.

Needed for #10647

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring