vinyl: support PP/NP iterator type
A PP/NP iterator basically acts exactly as LE/GE with the adjusted
search key so to enable support for these iterator types in vinyl,
all we need to do is to reuse prepare_start_prefix_iterator() when
starting a search. To achieve that, we also move the iterator type
normalization, which is currently performed by vy_read_iterator to
the vinyl_iterator wrapper.

Closes #10647

NO_DOC=the feature hasn't been documented yet;
       the original doc request will be updated