upgrade: add 2.11.6 to the downgrade version list
The `downgrade_versions.yml` workflow would complain (when we tag the
next release) if I don't do that. See #8319 for details and #10344 for
further thoughts.

NO_DOC=This is neither a new functionality, not a bugfix. It just
       reflects the new list of the released versions.
NO_CHANGELOG=see NO_DOC
NO_TEST=At least, the `downgrade_versions.yml` workflow verifies that
        this version is present in the list.