upgrade: add missing 2.11.6 downgrade version
The 2.11.6 version was recently released, but I forgot to update the
version list before the release. Anyway, let's fix it afterwards for the
sake of consistency.

This commit aims only the `release/2.11` branch. It can be considered as
a backport of the similar change that lands to 3.x branches, see #11198,
but the motivation to do the change is different.

NO_DOC=This is neither a new functionality, not a bugfix. It just
       reflects the new list of the released versions.
NO_CHANGELOG=see NO_DOC
NO_TEST=At least, the `downgrade_versions.yml` workflow verifies that
        this version is present in the list.