ci: verify the downgrade list before the release
There is a CI workflow that verifies correctness of the list of versions
to where the downgrade is possible, see commit 6d8563472785 ("ci: add
workflow to check downgrade versions").

However, this check is run only on push of a tag that marks the new
release. It is too late, because other CI workflows that deploy source
tarball and built packages are run by pushing this tag too. The uploads
are already in-progress or done, when a human sees that the downgrade
list check fails.

Let's add another criteria to run the check:

* the commit is `X.Y.Z-entrypoint-<...>`, and
* there is the `changelogs/X.Y.Z.md` file.

This way we can catch the problem while preparing the release notes,
*before the release*.

Follows up #8319

NO_DOC=development tools
NO_CHANGELOG=see NO_DOC
NO_TEST=see NO_DOC