tools: update release notes template
The given template comes from the time, when we had 1.10 and 2.x with
the old release policy (see [1]). In fact, almost every aspect of this
template is obsolete now.

For example, we have no stable and non-stable releases anymore: there
are pre-releases (marked with alpha, beta, rc keywords) and releases. We
don't need to say explicitly that the given release is stable.

Also, 1.10 is not LTS anymore (it is EOL) and 2.11 is not the latest
series.

The links needs an update too.

So, I decided to write these templates from scratch, but keep the idea:
give some release status information, highlight notable changes (for
feature releases), provide a few hints about the update procedure.

[1]: https://github.com/orgs/tarantool/discussions/6182

NO_DOC=dev tools
NO_CHANGELOG=same
NO_TEST=same