test: reduce iterations in tests
There are two tests that marked as "long": delete_insert.test.lua
and delete_replace_update.test.lua. Both tests cannot be executed
on AArch64 because execution time is exceeded timeout (320 sec).
Both tests runs a function many times, so these tests were
designed as a load tests, not a functional. Mixing load and
functional test is a bad idea, to allow executing these tests on
AArch64 the patch reduces a number of iterations and overall
times is reduced in 10 times. As a mitigation we have performance
tests that used as a load and performance tests.

Needed for the following patch.

Part of #10840

NO_CHANGELOG=testing
NO_DOC=testing