memtx: fix gap tracking with successor for func indexes
In memtx MVCC, we track gap reads right in successors if they are
available and then rely on this fact when providing transaction
isolation. The problem is, we don't support successors in func indexes,
so the isolation works incorrectly. The commit adds successor support
to single-key func indexes.

Follows up #11099

NO_CHANGELOG=later
NO_DOC=later