misc: add changelog and doc for functional keys in memtx mvcc
The commit adds changelog entries and doc request for the patchset
adding support of functional indexes to memtx MVCC. The first changelog
entry is about functional indexes themselves. All the issues are
considered as bugs and crashes, but since functional indexes were not
supported at all, we consider the patch as the feature, so the changelog
entry does not reference any issue. The second changelog is about
forbidding multikey indexes with MVCC - we should have done it much
earlier.

NO_TEST=no code changes

@TarantoolBot document

Title: Memtx MVCC unsupported index types

Tarantool with enabled memtx MVCC actually does not support multikey
indexes. Since Tarantool 3.4, an error will be thrown is one tries to
create such index. In earlier versions, it will lead to a crash or to a
panic - we should document it.

Note that if one wants to enable MVCC but it has multikey indexes, it
has to drop it and do a snapshot so that the definition of the index and
data inserted to it will disappear from snapshot and WAL completely.

Also, Tarantool with enabled memtx MVCC didn't support functional
indexes - it would lead to a crash as well, we should also document it.
Since Tarantool 3.4 functional indexes are supported (but only singe-key
ones).