ci: fix integration workflow
The `ubuntu-20.04 (focal)` image is no longer supported, so the build
using this image returns the following error:

| This is a scheduled Ubuntu 20.04 brownout. Ubuntu 20.04 LTS runner
| will be removed on 2025-04-01.

See [1] for the details.

This patch bumps the used version to the ubuntu-24.04 (noble).
Also, it forces the use of the branch version of the reusable_build
workflow to allow testing its changes.

The following modules should be fixed first before enabling the
integration with them back:
* tarantool/ddl
* tarantool/expirationd
* tarantool/tarantool-python

[1]: https://github.com/actions/runner-images/issues/11101

Relates to tarantool/ddl#130
Relates to tarantool/expirationd#176
Relates to tarantool/tarantool-python#334

NO_CHANGELOG=ci
NO_DOC=ci
NO_TEST=ci