config: add ability to run code before box.cfg
This commit adds 'tags' support for roles and app scripts.
Currently only 'preload' tag is supported. If this 'preload' tag is
specified for a role or a script, this role or script will be loaded
before the execution of `box.cfg` by the config module.

Closes #10182
Closes TNTP-227

@TarantoolBot document
Title: roles and scripts tags

Tags can be specified for roles and scripts (module or file in `app`
config section) in lua files. To add a tag to a source file,
add the following comment to the beginning of a file:

```lua
-- tags: preload
```

If several tags sections are provided all tags will be collected.

Currently only 'preload' tag is supported. The 'preload' tag makes it so
that the tagged file will be loaded (required) before the execution of
`box.cfg` by the config module.