config: add ability to run code before box.cfg
This commit adds 'tags' support for roles and app scripts.
Currently only 'early_load' tag is supported. If this 'early_load'
tag is set to `true` for a role or a script, this role or script will
be loaded before the execution of `box.cfg` by the config module.

Closes #10182
Closes TNTP-227

@TarantoolBot document
Title: roles and scripts tags

Tags can be specified for roles and scripts (module or file in `app`
config section) in lua files. To add a tag to a source file,
add the following comment to the beginning of a file (or after a
shebang):

```lua
-- early_load: true
```

Currently only 'early_load' tag is supported. Setting 'early_load' tag
makes it so that the tagged file will be loaded (required) before the
execution of `box.cfg` by the config module.