vinyl: implement index quantile method
We find the quantile key assuming that all keys are stored at the
deepest disk layer while newer layers contain incremental updates.
First, we walk over all index ranges overlapping with the target range
and estimate the number of tuples in them. Then we compute the offset of
the quantile key and walk over the index ranges again to find the key.

To achieve that, we need two helper functions:

 1. vy_run_estimate_stmt_count() estimates the number of tuples in
    the given run within the specified range. Basically, it finds
    the number of pages falling in the range, then scales down
    the total number of rows stored in the run proportionally,
    assuming that each page stores approximately the same number
    of rows.

 2. vy_run_estimate_key_at() finds a key at the given offset.
    Like vy_run_estimate_stmt_count(), this function assumes that
    each page stores approximately the same number of rows to
    find the page where the target key is located, then it returns
    the page min key, which should be a good enough approximation.

None of the functions accesses disk (yields).

Note that the target key may happen to be outside the target range, for
example, if the range is so small that there's no page fitting in it.
To follow the index:quantile() function protocol, we ignore such keys
(return NULL).

Also note that runs store extended keys so we need to strip extra parts
before returning the found key to the user.

Closes #11111

@TarantoolBot document
Title: Document `index:quantile()`

The new index method finds a quantile point in an indexed data range.
It takes the quantile level (0 < L < 1) and optionally the target range
boundaries and returns such a key that the ratio of tuples less than
the key equals L.

Example:

```lua
local s = box.schema.space.create('test')
s:create_index('pk')

for i = 1, 100 do
    s:insert({i * i, i})
end

-- Find the median in the whole index.
s.index.pk:quantile(0.5) -- returns {2601}

-- Find the 90th percentile among all keys < 1000
s.index.pk:quantile(0.9, {}, {1000}) -- returns {784}
```

The target range is defined as the intersection of the following read
requests:

```lua
s:select(begin_key, {iterator = 'ge'})
s:select(end_key, {iterator = 'lt'})
```

This means that using the empty key `{}` or `nil` for both `begin_key`
and `end_key` finds the quantile in the whole index. The function raises
an error if there can't possibly be any tuples in the target range, i.e.
if `key_begin` >= `key_end`.

The function returns `nil` if there are no tuples in the target
range or if it failed to find the quantile due to the limitations of
the storage engine, which are described below.

The function is implemented by memtx and vinyl tree indexes only.
In memtx the function returns the quantile point exactly. It has
the logarithmic complexity. It may return `nil` only if there are
no tuples in the target range. There's guaranteed to be a tuple
in the target range corresponding to the returned key.

The vinyl implementation returns the quantile point with a reasonable
degree of error for performance considerations. It has the logarithmic
complexity as well. There may or may not be a tuple in the target
range corresponding to the returned key. The function may return nil
even if the target range is not empty, in particular, if there are
no disk layers or if the range is smaller than the vinyl page size.

The function doesn't yield.

The function doesn't participate in transactions.

Like other index methods, `index:quantile()` can be used through
a space object as `space:quantile()`, in which case it works as
a shortcut for the primary index.

For more details, see the RFC document: https://github.com/orgs/tarantool/discussions/11194