build: fix install paths in Ubuntu Noble package
The tarantool-common package should install systemd service files and
the generator file into `/usr/lib/systemd` (not `/lib/systemd`) on
Ubuntu 24.04 Noble.

This commits fixes the paths in the same way as it was done in [1]. The
paths in the previous OS versions are not changed and a related test is
added.

Also, note that Debian Bookworm package should keep installing these
files into `/lib/systemd`. It was decided by the Debian Technical
Committee (quoted from [2]):

> Please stop moving individual packages' files from the root filesystem
> into /usr, at least for now.

The problem only affects 2.11 series, because 3.x doesn't ship systemd
service files.

[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1063699
[2]: https://lists.debian.org/debian-devel/2021/10/msg00190.html

Fixes #11234

NO_DOC=fixes a problem in a package, nothing to document
NO_CHANGELOG=since I'm going to rebuild 2.11.6 packages, then the
             problem will not be visible for the end user
NO_TEST=the test is added out of the usual testing infrastructure,
        because it verifies a package content, not a behavior of the
        executable file