ci: verify that Debian package has systemd service
The next commit fixes a problem related to installation paths of systemd
service files. This commit adds a regression test for it.

Part of #-WIP-11234

NO_DOC=a new CI check is added, no code or behavior changes
NO_CHANGELOG=same
NO_TEST=the CI check is added, it is in fact a test