config: don't choose learners as bootstrap leaders
This patch makes Tarantool configured in `replication.failover =
'supervised'` mode choose instances marked as learners for the failover
coordinator as a bootstrap leader only when there are no other
candidates.

Marking instance as a learner means it shouldn't be appointed as a
master by the failover. Thus, the user would expect for the learners not
to become RW during the bootstrap too.

Example.
```
replication:
  failover: supervised

failover:
  replicasets:
    r-001:
      learners:
        - i-001

groups:
  g-001:
    replicasets:
      r-001:
        instances:
          i-001: {}
          i-002: {}
```
`i-002` will be chosen as a replicaset bootstrap leader.

Instances marked as learners can still become RW if there are no other
candidates for bootstrap leaders. In this case the warning is issued
since it's likely to be an error in the configuration.

Closes #10842

NO_DOC=see tarantool/doc#4993