config: don't choose learners as bootstrap leaders
This patch makes Tarantool configured in `replication.failover =
'supervised'` mode not choose instances marked as learners for the
failover coordinator as a bootstrap leader.

Marking instance as a learner means it shouldn't be appointed as a
master by the failover. Thus, the user would expect for the learners not
to become RW during the bootstrap too.

Example.
```
replication:
  failover: supervised

failover:
  replicasets:
    r-001:
      learners:
        - i-001

groups:
  g-001:
    replicasets:
      r-001:
        instances:
          i-001: {}
          i-002: {}
```
`i-002` will be chosen as a replicaset bootstrap leader.

If all non-anonymous instances are marked as learners the initial
cluster bootstrap fails since every instance starts in RO and none
of them might become the leader.

If user applies a config where all non-anonymous instances marked as
learners an alert is issued since this config is likely to be wrong.

```
log.lua:74 W> box_cfg.apply: cannot determine a bootstrap leader based
on the configuration. Make sure learners are properly configured
```

Closes #10842

NO_DOC=see tarantool/doc#4993