memtx: implement the switch_to_ro callback
The `switch_to_ro` callback was implemented for memtx engine similar
to one we use in vinyl. We just abort all transactions so it is now
impossible for a fiber to awake and commit after switching the node to
read-only. Also more explicit error messages were implemented for both
memtx and vinyl to address that situation.

Fixes #9937
NO_DOC=bugfix