box: introduce `txn_synchro_timeout` config option
Fixed a bug when with the new `replication_synchro_timeout` behavior
(`compat.replication_synchro_timeout = 'new'`) a user fiber could hang
indefinitely if for some reason the quorum could not respond with an ACK
to a synchronous transaction.

Closes #10181

@TarantoolBot document
Title: new config option: 'txn_synchro_timeout'
Product: Tarantool

Root document: New page - https://www.tarantool.io/ru/doc/latest/reference/configuration/#txn-synchro-timeout

A new configuration option has been introduced - `txn_synchro_timeout`.
This timeout starts running at the moment when a synchronous transaction
starts to collect a quorum. If the timeout expires but the transaction
still has not reached a quorum, the user fiber is detached from the
transaction - the user got an error in this fiber 'Quorum collection for
a synchronous transaction is timed out. The transaction is detached from
this fiber and continues waiting for quorum'. The transaction is not
rolled back, it will wait as long as it takes to gather a quorum.

Option is only available with new `replication_synchro_timeout` behavior.
If `compat.replication_synchro_timeout = 'old'` user cannot set
non-default value in `txn_synchro_timeout`, they will get an error
"Incorrect value for option 'txn_synchro_timeout': option is disabled if
compat option `replication_synchro_timeout` is set to 'old'"
when trying to do so.