box: fix incorrect hashing of sequential JSON fields
Previously, if JSON fields were in a row in memory, the
'tuple_field_row()' function was used, which was not able to
correctly navigate through the nested JSON structure, which
ed to incorrect indexing in hash indexes. Now, if there are
JSON paths, 'tuple_field_raw_by_part()' is always called for
correct access to nested data, regardless of their sequence.

Closes https://github.com/tarantool/tarantool/issues/11014

NO_DOC=bug fix