box: check ability to compress a field regardless of space upgrade
We used to postpone all format checks during a space upgrade until the
upgrade starts to run. The idea was that the old tuples might not
conform to the new format until they are converted by the upgrade func.

Actually, this skipped some checks that should be run anyways, like
making sure we are not turning on compression for an indexed field.

Fix this.

Needed for tarantool/tarantool-ee#1166

NO_CHANGELOG=ee
NO_DOC=bugfix
NO_TEST=ee