config: call roles `stop` callback on shutdown
This commit makes it so the `stop` callbacks are called for all
registered roles, in the reverse order of roles startup.

Closed #10795
Closes TNTP-667

@TarantoolBot document
Title: roles are stopped on graceful shutdown

From now on during the graceful shutdown process `stop` callback for all
registered roles are called in the reverse order of roles startup.