memtx: introduce the O(n) SK build
Introduced a new `dump_primary_key` memtx index read view method. It's
meant to be defined for any index that can be used as a primary one
(currently TREE and HASH). The function is used in order to dump the
primary key tuples into the sort data file.

The new method is introduced in order to simplify the new .sortdata
file creation procedure: this way it can be separated from the main
.snap creation with smaller performance impact.

Introduced a new `dump_sort_data` memtx index read view method. It's
meant to be defined for any secondary key that can be sorted using the
new O(n) algorithm (currently, only TREE index supports it). This one
has to be an index-specific (along with the following one) in order to
provide the expected recovery speed-up.

Introduced a new `build_presorted` memtx index method using the sort
data along with the old2new tuple pointer map created from the primary
key to provide accelerated recovery.

A new entity introduced: memtx sort data file. This is a regular XLOG
file containing the data required to perform recovery using the O(n)
SK sorting algorithm. The file has `.sortdata` extension and garbage
collected the same way as regular `.snap` files.

The new `memtx_sort_data_enabled` config option controls whether to
write the sort data file on `box.snapshot()` or use it during recovery.
The option can be changed dynamically, so user can write checkpoints
with or without the new data regardless of had the data been used on
current instance recovery or had not.

But there are conditions when the sort data can not be used during
recovery:
0. The space is a system space (these generally don't benefit from the
   new SK sorting algorithm).
1. The `_index` space has no `before_replace` triggers (so we avoid
   the sort data invalidation of an index because the index had been
   altered by a `before_replace` trigger).
2. The space itself has no `before_replace` triggers (the new SK sort
   approach does not account this).

Of course, the new approach is also not used during forced recovery.

@TarantoolBot document
Title: Document the MemTX's new secondary key build approach
Since: 3.4.1

Now it's possible to sort secondary keys using a new O(n) sorting
algorithm using an additional data written into the snapshot. The
feature can be enabled with a new `memtx_sort_data_enabled` option
in `box.cfg` or `memtx.sort_data_enabled` in the instance config.

With the option set to `true`, additional data is saved into a
separated `.sortdata` file on snapshot creation and the data is
used on recovery. The default value is `false`: the behavior has
to be turned on by the user if required.

The option is changeable in runtime. So one can, for example, enable
it on recovery to use the new SK sorting approach, but disable it
before creating a new snapshot, so only `.snap` file will be created
and the O(n) SK sort will not be available for the snapshot.

The performance impact of the option depends on the persistent
storage read/write speed and the amount of tuples and secondary
keys in spaces (the more the better the new approach is). Also,
the approach only utilizes a single CPU core.

As a downside, it creates an additional footprint overhead on
recovery time (up to ~45 bytes per tuple) and overhead during
the snapshot creation (since we have to write the sort data to
the persistent storage).