config: add API to test if role is set for instance
New `has_role`, `is_router` and `is_storage` methods added to the config
module.

Closes #11288
Closes TNTP-969

@TarantoolBot document
Title: config module `has_role`, `is_router` and `is_storage` methods

`has_role(<role_name>, {instance = <instance_name})` returns true if
the instance with the name <instance_name> has role <role_name> enabled
in the current config, false if not. Second argument is optional, and
if not provided the check will be done for the instance the method is
called on.

`is_router({instance = <instance_name})` returns true if the instance
with name <instance_name> is a vshard router, according to the current
config, false if not. Argument is optional, and if not provided the
check will be done for the instance the method is called on.

`is_storage({instance = <instance_name})` returns true if the instance
with name <instance_name> is a vshard storage, according to the current
config, false if not. Argument is optional, and if not provided the
check will be done for the instance the method is called on.